+++ tests/cmake/find_package/legacy_transport_defaults.cpp
@@ -7,7 +7,7 @@ namespace {
 
 class LegacyTransportImplementation final : public orpheus::ITransportController {
 public:
-  orpheus::SessionGraphError startClip(orpheus::ClipHandle) override {
+  orpheus::SessionGraphError startClip(orpheus::ClipHandle, orpheus::StartRequestTag = 0) override {
     return ok;
   }
   orpheus::SessionGraphError stopClip(orpheus::ClipHandle) override {